agentic-ai-solution-backend/pyproject.toml
[tool.poetry]
name = "agentic-ai-solution-backend"
version = "0.1.0"
description = "FastAPI backend with LangGraph agents for Agentic AI Solution"
authors = ["Development Team <dev@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
pydantic = "^2.7.0"
httpx = "^0.27.0"
anthropic = "^0.26.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.0"

# Optional LLM providers
google-generativeai = { version = "^0.4.0", optional = true }
openai = { version = "^1.0.0", optional = true }

[tool.poetry.extras]
google = ["google-generativeai"]
openai = ["openai"]
all-llms = ["google-generativeai", "openai"]

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
